More responsive and mobile friendly landing page   - Move all resources to local (makes visualizing the page localy much easier)   - Merge `pdoc3` and `pages` into `pages`   - Change docs to be generated under `public/documentation`
