+++ pages/index.html
@@ -2,28 +2,7 @@
 <html lang="en">
 
 <head>
-    <!--
-     | Copyright (c) Nekokatt 2020
-     |
-     | Permission is hereby granted, free of charge, to any person obtaining a copy
-     | of this software and associated documentation files (the "Software"), to deal
-     | in the Software without restriction, including without limitation the rights
-     | to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-     | copies of the Software, and to permit persons to whom the Software is
-     | furnished to do so, subject to the following conditions:
-     |
-     | The above copyright notice and this permission notice shall be included in all
-     | copies or substantial portions of the Software.
-     |
-     | THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-     | IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-     | FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-     | AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-     | LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-     | OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-     | SOFTWARE.
-     !-->
-    <link rel="shortcut icon" type="image/png" href="https://hikari-py.github.io/hikari/logo.png" />
+    <link rel="shortcut icon" type="image/png" href="logo.png" />
     <meta property="og:title" content="Hikari.py" />
     <meta property="og:type" content="website" />
     <meta property="og:image" content="https://hikari-py.github.io/hikari/logo.png" />
@@ -127,7 +106,7 @@
             <div class="inner">
                 <ul class="nav justify-content-center">
                     <li class="nav-item"><a class="nav-link active" href="#">Home</a></li>
-                    <li class="nav-item"><a class="nav-link" href="hikari/index.html">Documentation</a></li>
+                    <li class="nav-item"><a class="nav-link" href="documentation/index.html">Documentation</a></li>
                     <li class="nav-item"><a class="nav-link" href="https://github.com/hikari-py/hikari">GitHub</a></li>
                     <li class="nav-item"><a class="nav-link" href="https://pypi.org/project/hikari">PyPI</a></li>
                 </ul>
@@ -135,10 +114,7 @@
         </header>
 
         <main role="main" class="inner cover">
-            <h1 class="display-1">
-                <small>pip install --pre</small><br />
-                hikari
-            </h1>
+            <h1 class="display-1"><small>pip install</small><br />hikari</h1>
             <p class="lead">A new, powerful, static-typed Python API for writing Discord bots.</p>
             <p>
                 This API is still in a alpha state, and is a <strong>work in progress</strong>! Features may change
@@ -171,7 +147,7 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
         integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
         crossorigin="anonymous"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.0.1/highlight.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.0.1/highlight.min.js"></script>
     <script>
         hljs.initHighlightingOnLoad();
         $(document).ready(function () {
